+++ App/Controllers/ProfileController.php
@@ -56,15 +56,19 @@ class ProfileController extends AControllerBase
     {
         $req = $this->request();
         $user_id = $req->getValue('user_id');
+        $errors = $req->getValue('errors');
         $user_data = Data::getAll("`user` = ?", [$user_id]);
         $folders = Folder::getAll("`owner` = ?", [$user_id]);
-        return $this->html(['user_data' => $user_data, 'folders' => $folders]);
+        return $this->html(['user_id' => $user_id, 'user_data' => $user_data, 'folders' => $folders, 'errors' => $errors]);
     }
 
     public function profileFolder()
     {
         $req = $this->request();
         $folder = Folder::getOne($req->getValue('folder'));
+        if ($folder == null) {
+            return new RedirectResponse($this->url("folder.error", ['message' => "Folder doesn't exist"]));
+        }
         $folder_data = $folder->getAllFromFolder();
         return $this->html(['folder_data' => $folder_data, 'folder' => $folder]);
     }
@@ -77,7 +81,7 @@ class ProfileController extends AControllerBase
         $profile = Profile::getAll("`user` = ?", [$user_id])[0];
         if (is_null($req->getValue('submit_profile')))
         {
-            return $this->html(['user_id' => $user_id, 'username' => $username, 'profile' => $profile]);
+            return $this->html(['user_id' => $user_id, 'username' => $username, 'profile' => $profile, 'errors' => ""]);
         }
 
         $new_username = $req->getValue('username');
@@ -91,7 +95,7 @@ class ProfileController extends AControllerBase
         else if ($new_username != $username)
         {
             $fetched_user = User::getAll("`username` = ?", [$new_username]);
-            if ($fetched_user && $fetched_user[0]->getUsername() == $username && $fetched_user[0]->getId() != $user_id)
+            if ($fetched_user && $fetched_user[0]->getUsername() == $new_username && $fetched_user[0]->getId() != $user_id)
             {
                 return $this->html(['user_id' => $user_id, 'username' => $username, 'profile' => $profile, 'errors' => 'User already exists']);
             }
@@ -101,50 +105,16 @@ class ProfileController extends AControllerBase
             $username = $new_user->getUsername();
         }
         $profile->setDescription($new_desc);
-        $new_file_location = $this::$PICTURE_DIR.rand(0, 255).'-'.$new_profile_pic_file['name'];
-        move_uploaded_file($new_profile_pic_file['tmp_name'], $new_file_location);
-        unlink($profile->getProfilePic());
-        $profile->setProfilePic($new_file_location);
-        $profile->save();
-        return $this->html(['user_id' => $user_id, 'username' => $username, 'profile' => $profile]);
-    }
-
-    public function editAction() : Response
-    {
-        $req = $this->request();
-        $user_id = $req->getValue('user_id');
-        $username = User::getAll("`id` = ?", [$user_id])[0]->getUsername();
-        $profile = Profile::getAll("`user` = ?", [$user_id])[0];
-
-        $new_username = $req->getValue('username');
-        $new_desc = $req->getValue('description');
-        $new_profile_pic_file = $req->getFiles()['profile_pic'];
-
-        if (is_null($new_username))
-        {
-            return $this->html(['user_id' => $user_id, 'username' => $username, 'profile' => $profile, 'errors' => 'Username cannot be empty']);
-        }
-        else if ($new_username != $username)
-        {
-            $fetched_user = User::getAll("`username` = ?", [$new_username]);
-            if ($fetched_user && $fetched_user[0]->getUsername() == $username && $fetched_user[0]->getId() != $user_id)
-            {
-                return $this->html(['user_id' => $user_id, 'username' => $username, 'profile' => $profile, 'errors' => 'User already exists']);
-            }
-            $new_user = User::getOne($user_id);
-            $new_user->setUsername($new_username);
-            $new_user->save();
-            $username = $new_user->getUsername();
-        }
-        $profile->setDescription($new_desc);
-        if ($new_profile_pic_file['name'] != "") {
+        if ($new_profile_pic_file['name'] != null) {
             $new_file_location = $this::$PICTURE_DIR.rand(0, 255).'-'.$new_profile_pic_file['name'];
             move_uploaded_file($new_profile_pic_file['tmp_name'], $new_file_location);
-            unlink($profile->getProfilePic());
+            if ($profile->getProfilePic() != null || $profile->getProfilePic() != "") {
+                unlink($profile->getProfilePic());
+            }
             $profile->setProfilePic($new_file_location);
         }
         $profile->save();
-        return new RedirectResponse($this->url("profile",  ['id' => $user_id]));
+        return $this->html(['user_id' => $user_id, 'username' => $username, 'profile' => $profile, 'errors' => ""]);
     }
 
     public static function addProfile(User $user)
