+++ App/Controllers/ReportController.php
@@ -4,6 +4,7 @@ namespace App\Controllers;
 
 use App\Core\AControllerBase;
 use App\Core\Responses\Response;
+use App\Models\Data;
 use App\Models\ReportType;
 use App\Models\Report;
 use App\Models\User;
@@ -46,6 +47,18 @@ class ReportController extends AControllerBase
         $data_id = $req->getValue('data_id');
         $report_type = $req->getValue('report_type');
 
+        if (User::getOne($user_id) == null) {
+            return $this->redirect($this->url("report.error", ["message" => "User doesn't exist"]));
+        }
+
+        if (Data::getOne($data_id) == null) {
+            return $this->redirect($this->url("report.error", ["message" => "Data doesn't exist"]));
+        }
+
+        if (ReportType::getOne($report_type) == null) {
+            return $this->redirect($this->url("report.error", ["message" => "Invalid report type"]));
+        }
+
         $report = new Report();
         $report->setUser($user_id);
         $report->setData($data_id);
@@ -60,8 +73,17 @@ class ReportController extends AControllerBase
         $req = $this->request();
         $report_id = $req->getValue('report_id');
         $report = Report::getOne($report_id);
+        if ($report == null) {
+            return $this->redirect($this->url("report.error", ["message" => "Report doesn't exists"]));
+        }
         $report->delete();
-
         return $this->redirect($this->url("adm.adminReports"));
+    }
+
+    public function error()
+    {
+        $req = $this->request();
+        $message = $req->getValue('message');
+        return $this->html(["message" => $message]);
     }
 }
